@@ -7,30 +7,31 @@
   "private": true,
   "proxy": "http://localhost:3000/",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.8",
-    "@fortawesome/free-regular-svg-icons": "^5.5.0",
-    "@fortawesome/free-solid-svg-icons": "^5.5.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.12",
+    "@fortawesome/free-regular-svg-icons": "^5.6.3",
+    "@fortawesome/free-solid-svg-icons": "^5.6.3",
     "@fortawesome/react-fontawesome": "^0.1.3",
-    "@types/jest": "^23.3.9",
-    "@types/moment-timezone": "^0.5.9",
-    "@types/node": "^10.12.2",
-    "@types/react": "^16.4.18",
-    "@types/react-dom": "^16.0.9",
+    "@types/jest": "^23.3.12",
+    "@types/moment-timezone": "^0.5.10",
+    "@types/node": "^10.12.18",
+    "@types/query-string": "^6.2.0",
+    "@types/react": "^16.7.18",
+    "@types/react-dom": "^16.0.11",
     "@types/react-router-dom": "^4.3.1",
     "axios": "^0.18.0",
     "jwt-decode": "^2.2.0",
     "moment-timezone": "^0.5.23",
-    "node-sass": "^4.9.4",
+    "node-sass": "^4.11.0",
     "query-string": "^6.2.0",
-    "react": "^16.6.0",
+    "react": "^16.7.0",
     "react-cookies": "^0.1.0",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.7.0",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1",
-    "react-select": "^2.1.1",
-    "tslint": "^5.11.0",
+    "react-scripts": "2.1.3",
+    "react-select": "^2.2.0",
+    "tslint": "^5.12.1",
     "tslint-react": "^3.6.0",
-    "typescript": "^3.1.6"
+    "typescript": "^3.2.2"
   },
   "scripts": {
     "build": "react-scripts build",
